// webapp/resources/sap/esh/search/ui/SearchResultTableColumnType.d.ts
declare module "sap/esh/search/ui/SearchResultTableColumnType" {
    /*!
     * SAPUI5
     * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
     *
     */
    enum TableColumnType {
        TITLE = "TITLE",
        TITLE_DESCRIPTION = "TITLE_DESCRIPTION",
        DETAIL = "DETAIL",
        RELATED_APPS = "RELATED_APPS",
        EXTEND = "EXTEND"
    }
}
//# sourceMappingURL=SearchResultTableColumnType.d.ts.map