// webapp/resources/sap/esh/search/ui/UIEvents.ts
/*!
 * SAPUI5
 * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
 * 
 */
enum UIEvents {
    ESHSearchFinished = "ESHSearchFinished",
    ESHSearchStarted = "ESHSearchStarted",
    ESHSearchLayoutChanged = "ESHSearchLayoutChanged",
    ESHSelectionChanged = "ESHSelectionChanged",
    // ESHResultViewSelectionChanged = "ESHResultViewSelectionChanged",  ToDo, activate as of 1.142
    ESHResultViewTypeChanged = "ESHResultViewTypeChanged",
    ESHShowResultDetail = "ESHShowResultDetail",
}

export default UIEvents;