// webapp/resources/sap/esh/search/ui/SelectionMode.d.ts
declare module "sap/esh/search/ui/SelectionMode" {
    /*!
     * SAPUI5
     * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
     *
     */
    enum SelectionMode {
        None = "None",
        OneItem = "OneItem",
        MultipleItems = "MultipleItems"
    }
}
//# sourceMappingURL=SelectionMode.d.ts.map